Turns OCR into an object
This commit intends to turn the ocr function handler into an object that can be instantiated by other functions, this will allow us to have unique interpreters in parallel.

One use case for that is being able to interpret both portuguese and english text at the same time in two different instances of the OCR object.

Signed-off-by: gabrieldrs <gabriel@sezefredo.com.br>