Adds ESLint to project
This commit adds a set of ESLint rules to the project with the intention
to improve code quality and code standards. This also applies the defined
rules to the already created files in the projet.

ESlint is available as a plugin for many editors like Visual Studio Code
or Atom, these plugins will highlight errors/warnings while you develop,
so it's easy to track and fix. It's highly recommended to use the ESLint
plugin.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>