Adds Cloudant dependency and new env vars to .env
I added a file called .env-example that will act as a boilerplate for
newcomers, this file lists all environment variables we use in the
application.

Instructions on how to use it were added to the README file, but
basically, the developer will have to create a copy of .env-example called .env
and fill in the environment variables values. dotenv extension will automatically
load these environment variables when we run the application.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>