Adds file system client to the API
In this commit I'm starting to build wrappers for our third party integrations.
The first one being implemented is the Box API, for that, I created a FSClient object that will wrap access to Box API into a simple CRUD object.
The intention here is to make it easier to change our file system solution, if needed.

As a v1, I created 3 functions:

 - uploadMenu: Specific setup to uplaod our menus to the right folder, without the need to specify paths, this will have the knowledge of where the menus are being stored.
 - upload: Generic upload function, if we need to upload something else, we can use this function to do so.
 - getFolderId: This is a bit too specific to box, so we may want to make this private later on, but this is a function intended to get the id of a specific folder, so we can upload files to it.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>