Creates a multer configurator module
MulterConfig will hold a list of configs for different resources that we will want to upload.
Initially it holds the configuration to upload menus.
When we configure the "/post" route for uploading menus, we will call this in order to generate the middleware following the configrations defined in this module.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>