Import required configurations to the "menus" resource router
In this commit I use the previously created MulterConfig module in order to create the upload configurator for our menus.post('/') route.
An edit was necessary in the menus error handler. As noted in the code, file size errors are generated directly by multer/busboy, and because of that it was a bit hard to define a specific error status for this case, so, in order to accomplish this, I had to add an error code check to the "menus" error handler to manually set the error status I wanted when a file too large is being sent. We may be able to improve this later, but I didn't find this to be a blocker.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>