Adds an optional db parameter to DBClient
In our first implementation, DBClient had to receive a db parameter on all
of its methods. This commit adds an optional db parameter to its
constructor so if it's defined, a db parameter is not necessary in method
calls.

This commit also adapts the menu "insert" route to use this new
functionality.

A helper function "validateDB" was created in order to check if a db was
specified in a method call, it can be used in all methods of DBClient.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>