Closes #19 - Implements error messages module
We started to have the same error message written in many different
places, because of that we decided to implement an error messages module
in order to unify error messages and promote reuse.

The module exports only functions, I decided to export a function even
when no parameter is necessary for standardization.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>