Add menu schema
The api/lib/schemas/menu.js contains rules with for how to specifie a menu
document. Considering that the ocr and file will be generated if the I/O
ops went well, they don't need to be required because the user can update
them. The _id and _rev are generated automatically by the Cloudant, so just
the restaurant object must be required.
The index.js will be usefull in cases that we will implement more than one
schema.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>