Closes #20 - Add JSONValidator class
This class starts getting the current schemas from the schemas folder,
saving them in the validSchemas variable.
If the user uses isValid function, this function get the schema passed to see
if it is a valid schema and if it is make the verification of the document
with the schema parameter.
The getSchema function is used to print the schema passed as it's parameter.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>