Add getOne menu route
This route can be used by a get request to /api/v1/menus/:id
It uses getById dependency from the DBClient to find the
requested id in it's parameter. If the DBClient returns a
'missing' error.message the status code to the client is 404,
otherwise is going to be 500.
There is no need to specifie the typeof the id parameter in the route
and it's maximum size or invalid characters, beacause it is already
defined as a string in the route - and cannot be undefined - and
the Cloudant db does not have limitations about speacial character
or limit size to the id key value.

Signed-off-by: Gustavo Porto Guedes gustavo.guedes@fatec.sp.gov.br