Closes #24 - Implements new "sanitize" method in JSONValidator
This commit adds a "sanitize" function in our JSONValidator. this method receives a json document and a schema name and returns a json document containing only fields defined in the schema.

A helper function "sanitizeList" is also implemented that receives a list and a schema name, and runs "sanitize" for each item of the list, passing the schema name.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>