Closes #23 - Adds JSON Validation steps to DBClient
This commit adds required validation steps to the DBClient object, they are being handled inside the Promise so we can catch errors in the Promise chain. Some checks were also moved inside the Promises, so we can also benefit on the promise chain on those.

This commit also adds a new label to menu schema, "menus", that lets us use the "menus" db name to get this schema.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>