Implement delete menu route
This route receives an menu id as it parameter.
There is no need to make an additional validation of this parameter
because it comes already as 'string' and with length more than 0.

It uses the getById function from the DBClient.js to retrive the
actual _rev and fileId of this document. If it catchs an error of
'missing' or 'deleted' means that the document wasn't found, so it
falls in our lib error for DB_DOC_NOT_FOUND. Any error that don't
qualify on the previous errors are considered INTERNAL_SERVER_ERROR.

With the document deleted on the database, it's start the fsClient
function to delete the file asynchronously with the success response
to the user.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>