Add paging to get menus route
This route is available with a get request to /api/v1/menus?from=?&size=?
It's make a verification with the query params passed in the request.
The params object receives the req.query.from and the req.query.size
if they are defined, they are converted to Number to verify after if
they are integer, otherwise the route throws an error.

After pass to the route check the params are passed to the DbClient get
function that has it's own verification of params too.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>