Add OCR lib
This lib takes care off the most important functionality of our
App, that is to convert PDF scans to text.

When you initiate a constructor you can specifie the language, psm
and where your tesseract software is installed in your operating
system. By default, the language is set to portuguese.

To use the OCR in your PDF file, you must use the getText function
that uses a private function called PDFtoImage to convert all your
PDF pages to images and then run tesseract to convert the images in
a readable string by the ImagesToText private function.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>