Closes #26 - Add Tesseract to the create pipeline
Was added on the final of create pipeline the tesseract.getText to
update the menu document on the menus db.

Was updated the MulterConfig to use the multer.diskStorage strategy, to
allow filename changes on the document sent by the user. For more information,
you can access the [Multer Documentation](https://www.npmjs.com/package/multer).

Finally, was removed the deprecated route '/v1/convert/ocr' from the server.js.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>